pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "flagrant"
version = "0.1.0"
description = "Local AI code reviewer for pre-commit checks."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Flagrant" },
]
keywords = ["code-review", "ai", "cli", "linting", "developer-tools"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "gitpython>=3.1.0",
    "anthropic>=0.40.0",
    "openai>=1.0.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
flagrant = "flagrant.main:app"

[project.urls]
Homepage = "https://github.com/flagrant/flagrant"
Repository = "https://github.com/flagrant/flagrant"
